Run ARDY on Apple Silicon (MPS)
- ardy.tools.get_default_device(): cuda > mps > cpu, overridable with ARDY_DEVICE; used by
  load_text_encoder, the LLM2Vec encoder, generate.py, run_demo.py and the demo's model tab
  (new "mps / float16" text-encoder options). seed_everything also seeds the MPS generator.
- Skeleton buffers (joints.p etc.) are float64 on disk; cast to float32 since MPS has no f64.
- The text encoder runs in float16 on MPS (bf16 has no native math on M1/M2, ~2x slower).
- LLM2Vec.from_pretrained: pass dtype= (the transformers 5 name for torch_dtype).
- README: Apple Silicon setup section.

The default 16 GB text encoder still needs a 32 GB+ Mac (or a remote encoder); the int4
checkpoint path for 16 GB machines is a separate change on top of this one.

+++ ardy/skeleton/base.py
@@ -69,16 +69,17 @@ class SkeletonBase(torch.nn.Module):
 
         if load and folder is not None:
             pfolder = Path(folder)
-            neutral_joints = torch.load(pfolder / "joints.p").squeeze()
+            # Cast to float32: the pickled joints are float64, which MPS does not support.
+            neutral_joints = torch.load(pfolder / "joints.p").squeeze().float()
             self.register_buffer("neutral_joints", neutral_joints, persistent=False)
 
             if (pfolder / "bvh_joints.p").exists():
-                bvh_neutral_joints = torch.load(pfolder / "bvh_joints.p").squeeze()
+                bvh_neutral_joints = torch.load(pfolder / "bvh_joints.p").squeeze().float()
                 self.register_buffer("bvh_neutral_joints", bvh_neutral_joints, persistent=False)
 
             global_offset_path = pfolder / "standard_t_pose_global_offsets_rots.p"
             if global_offset_path.exists():
-                global_rot_offsets = torch.load(global_offset_path).squeeze()
+                global_rot_offsets = torch.load(global_offset_path).squeeze().float()
                 self.register_buffer("global_rot_offsets", global_rot_offsets, persistent=False)
 
         self.bone_order_names = [x for x, y in self.bone_order_names_with_parents]
